Feature/enhancements and bugfixes by @Mawiguk0

+++ src/app/Http/Controllers/Admin/Events/ParticipantsController.php
@@ -63,8 +63,17 @@ class ParticipantsController extends Controller
      */
     public function signIn(Event $event, EventParticipant $participant)
     {
+        if ($participant->event->slug != $event->slug)
+        {
+            Session::flash('alert-danger', 'The selected participant does not belong to the selected event!');
+            return Redirect::to('admin/events/' . $event->slug . '/participants/'); 
+        }
         if ($participant->ticket && $participant->purchase->status != "Success") {
             Session::flash('alert-danger', 'Cannot sign in Participant because the payment is not completed!');
+            return Redirect::to('admin/events/' . $event->slug . '/participants/' . $participant->id);
+        }
+        if ($participant->revoked) {
+            Session::flash('alert-danger', 'Cannot sign in a revoked Participant!');
             return Redirect::to('admin/events/' . $event->slug . '/participants/' . $participant->id);
         }
         if (!$participant->setSignIn()) {
@@ -72,11 +81,16 @@ class ParticipantsController extends Controller
             return Redirect::to('admin/events/' . $event->slug . '/participants/' . $participant->id);
         }
         Session::flash('alert-success', 'Participant Signed in!');
-        return Redirect::to('admin/events/' . $event->slug . '/participants/' . $participant->id);
+        return Redirect::to('admin/events/' . $event->slug . '/participants/');
     }
 
     public function transfer(Event $event, EventParticipant $participant, Request $request)
     {
+        if ($participant->event->slug != $event->slug)
+        {
+            Session::flash('alert-danger', 'The selected participant does not belong to the selected event!');
+            return Redirect::to('admin/events/' . $event->slug . '/participants/'); 
+        }
         if ($participant->ticket && $participant->purchase->status != "Success") {
             Session::flash('alert-danger', 'Cannot sign in Participant because the payment is not completed!');
             return Redirect::to('admin/events/' . $event->slug . '/participants/' . $participant->id);
@@ -120,8 +134,39 @@ class ParticipantsController extends Controller
         return Redirect::to('admin/events/' . $event->slug . '/participants/');
     }
 
+    /**
+     * Sign out a single participant for the event
+     * @param  Event  $event
+     * @param  EventParticipant $participant
+     * @return View
+     */
+    public function signout(Event $event, EventParticipant $participant)
+    {   
+        if ($participant->event->slug != $event->slug)
+        {
+            Session::flash('alert-danger', 'The selected participant does not belong to the selected event!');
+            return Redirect::to('admin/events/' . $event->slug . '/participants/'); 
+        }
+        if ($participant->revoked) {
+            Session::flash('alert-danger', 'Cannot sign out a revoked Participant!');
+            return Redirect::to('admin/events/' . $event->slug . '/participants/' . $participant->id);
+        }
+        if (!$participant->setSignIn(false)) {
+            Session::flash('alert-danger', 'Cannot sign out Participant! '. $participant->name);
+            return Redirect::to('admin/events/' . $event->slug . '/participants/');
+        }
+
+        Session::flash('alert-success', 'Participant ' . $participant->name . ' signed out!');
+        return Redirect::to('admin/events/' . $event->slug . '/participants/');
+    }
+
     function revoke(Event $event, EventParticipant $participant)
     {
+        if ($participant->event->slug != $event->slug)
+        {
+            Session::flash('alert-danger', 'The selected participant does not belong to the selected event!');
+            return Redirect::to('admin/events/' . $event->slug . '/participants/'); 
+        }
         if (!$participant->setRevoked()) {
             Session::flash('alert-danger', 'Cannot revoke Participant!');
             return Redirect::to('admin/events/' . $event->slug . '/participants/' . $participant->id);
@@ -132,6 +177,11 @@ class ParticipantsController extends Controller
 
     function delete(Event $event, EventParticipant $participant)
     {
+        if ($participant->event->slug != $event->slug)
+        {
+            Session::flash('alert-danger', 'The selected participant does not belong to the selected event!');
+            return Redirect::to('admin/events/' . $event->slug . '/participants/'); 
+        }
         if (!$participant->delete()) {
             Session::flash('alert-danger', 'Cannot delete participant');
             return Redirect::to('admin/events/' . $event->slug . '/participants/' . $participant->id);
